Create launch page and userPage
Create an initial launch page that prompts user to register for an
account or login if they are already a registered user. Once the user
logs in they are taken to the user page where they have to options to
play on a networked game or local game, go to the How to play page or
Settings page.