Created Register Page and Registration Page
The register page and login page were created with their basic non
functional layout that corresponds with the look and feel of the other
pages making it consistent with the design of the whole application.
Revisions were made in the form of naming conventions and other minor
changes have been made as well.