A new way to handle the background image
I took the background image and put all the logic involving it into a
separate page. Instead of the background loading upon each page change,
it checks the time every second to see if it needs to change, and if
needed it'll fade into the next background.
The catch is that the game has to be opened from the new "Index.html"
file.