Triggered Events, Product Removal patch
Coded the logic for three not-randomly triggered events: one to remind
the player to move along, one to remove employees if they spend too much
time on one product, and one to remove a product if it just fails that
badly.
Also added a check in the removeProduct function to also remove the
patent for that product's category, if it's patented.