Audio Settings page and default music
I finished the audio settings page and included some default music, both
mp3 and ogg for total compatibility (a wav file of any of them would be
larger in file size than both the mp3 and the ogg versions of that same
file combined; it's nuts).
I also ran into an issue regarding users adding custom music, upon which
I'll elaborate on GitHub in a moment.