Animations, a bug, and UI things
Products are sorta transparent and blink when selected. Important events
are accompanied with a little flash. Products on the board point to the
center. Finally fixed a bug that caused products to not be removed at
times. Did some optimizing of the sound playing function. Muting buttons
look better. Did some realigning on the title page.