VI Refactoring, RandomEvent fixing,
Big change here: Actions by the computer are taken recursively now; this
not only allows for an easier handling of them and a reasonable amount
of pop-up closing, but now the AI can be paused by pressing the "Quit"
button.
RandomEvents are now better handled as well, and style changes were made
as needed.