alright ryan, have at it
search for the string [NW]. It's in comments where you come in. I've
pulled everything up so each of the 8 functions are able to be sent and
gotten over the wire and execute remotely. All you have to do is plug
them in so they can communicate with eachother. The game should
self-handle being online, though it's up to you to get the objects build
in each client's localStorage before the game starts. After that, as
long as things communicate, the game will run as normal. Look for [NW]
and read the instructions.